Refactor: Extract validation from TransferFnudsDialog
